Fix ESLint warning: separate AuthContext and AuthProvider for Fast Refresh compatibility
