+++ frontend/src/contexts/AuthProvider.tsx
@@ -1,0 +1,133 @@
+import React, { useState, useEffect, ReactNode } from 'react';
+import { AuthContext, type AuthContextType, type RegisterData } from './AuthContext';
+import api from '../services/api';
+
+interface AuthProviderProps {
+  children: ReactNode;
+}
+
+export const AuthProvider: React.FC<AuthProviderProps> = ({ children }) => {
+  const [user, setUser] = useState<AuthContextType['user']>(null);
+  const [isLoading, setIsLoading] = useState(true);
+
+  // Verificar se há token salvo no localStorage
+  useEffect(() => {
+    const checkAuth = async () => {
+      try {
+        const token = localStorage.getItem('accessToken');
+        if (token) {
+          // Configurar o token no axios
+          api.defaults.headers.common['Authorization'] = `Bearer ${token}`;
+          
+          // Verificar se o token ainda é válido
+          const response = await api.get('/auth/me');
+          setUser(response.data.data);
+        }
+      } catch (error) {
+        // Token inválido, limpar localStorage
+        localStorage.removeItem('accessToken');
+        localStorage.removeItem('refreshToken');
+        delete api.defaults.headers.common['Authorization'];
+      } finally {
+        setIsLoading(false);
+      }
+    };
+
+    checkAuth();
+  }, []);
+
+  const login = async (email: string, password: string) => {
+    try {
+      const response = await api.post('/auth/login', { email, password });
+      const { user, accessToken, refreshToken } = response.data.data;
+
+      // Salvar tokens no localStorage
+      localStorage.setItem('accessToken', accessToken);
+      localStorage.setItem('refreshToken', refreshToken);
+
+      // Configurar token no axios
+      api.defaults.headers.common['Authorization'] = `Bearer ${accessToken}`;
+
+      setUser(user);
+    } catch (error: any) {
+      throw new Error(error.response?.data?.message || 'Erro ao fazer login');
+    }
+  };
+
+  const logout = async () => {
+    try {
+      const token = localStorage.getItem('accessToken');
+      if (token) {
+        await api.post('/auth/logout', {}, {
+          headers: { Authorization: `Bearer ${token}` }
+        });
+      }
+    } catch (error) {
+      // Ignorar erros no logout
+      console.warn('Erro ao fazer logout:', error);
+    } finally {
+      // Limpar dados locais
+      localStorage.removeItem('accessToken');
+      localStorage.removeItem('refreshToken');
+      delete api.defaults.headers.common['Authorization'];
+      setUser(null);
+    }
+  };
+
+  const register = async (userData: RegisterData) => {
+    try {
+      const response = await api.post('/auth/register', userData);
+      const { user, accessToken, refreshToken } = response.data.data;
+
+      // Salvar tokens no localStorage
+      localStorage.setItem('accessToken', accessToken);
+      localStorage.setItem('refreshToken', refreshToken);
+
+      // Configurar token no axios
+      api.defaults.headers.common['Authorization'] = `Bearer ${accessToken}`;
+
+      setUser(user);
+    } catch (error: any) {
+      throw new Error(error.response?.data?.message || 'Erro ao registrar usuário');
+    }
+  };
+
+  const refreshToken = async () => {
+    try {
+      const refreshToken = localStorage.getItem('refreshToken');
+      if (!refreshToken) {
+        throw new Error('Refresh token não encontrado');
+      }
+
+      const response = await api.post('/auth/refresh', { refreshToken });
+      const { accessToken, refreshToken: newRefreshToken } = response.data.data;
+
+      // Atualizar tokens no localStorage
+      localStorage.setItem('accessToken', accessToken);
+      localStorage.setItem('refreshToken', newRefreshToken);
+
+      // Configurar novo token no axios
+      api.defaults.headers.common['Authorization'] = `Bearer ${accessToken}`;
+    } catch (error: any) {
+      // Se o refresh falhar, fazer logout
+      await logout();
+      throw new Error('Sessão expirada. Faça login novamente.');
+    }
+  };
+
+  const value: AuthContextType = {
+    user,
+    isAuthenticated: !!user,
+    isLoading,
+    login,
+    logout,
+    register,
+    refreshToken,
+  };
+
+  return (
+    <AuthContext.Provider value={value}>
+      {children}
+    </AuthContext.Provider>
+  );
+};
